baseline-broker/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.5
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
azure-cognitiveservices-speech==1.42.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
comm==0.2.3
command-not-found==0.3
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
DateTime==6.0
dbus-python==1.2.18
debugpy==1.8.17
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.3.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flash==1.0.3
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gp==1.0
gplearn==0.4.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==22.10.0
iniconfig==2.3.0
ipykernel==6.30.1
ipython==8.12.3
ipywidgets==8.1.7
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
learn==1.0.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
mysql-connector-python==9.6.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.4.6
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.0
opencv-python==4.12.0.88
openpyxl==3.1.5
overrides==7.7.0
p==1.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.47
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.27.3
psutil==7.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyGObject==3.42.1
pygsheets==2.0.6
PyJWT==2.3.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyproject_hooks==1.1.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pytest==9.0.3
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2024.1
PyYAML==5.4.1
pyzmq==26.0.3
questdb==2.0.3
questdb-connect==1.0.11
queuelib==1.7.0
redis==7.4.0
referencing==0.35.1
regex==2025.11.3
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.15.3
Scrapy==2.11.2
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
service-identity==24.1.0
simplejson==3.20.2
six==1.16.0
slackclient==2.9.4
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.14.0
systemd-python==234
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tldextract==5.1.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
Twisted==24.3.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.13.0+ds
w3lib==2.2.1
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.14
yahoo-finance==1.4.0
yarg==0.1.9
yarl==1.9.4
ystockquote==0.2.5
zipp==1.0.0
zope.interface==6.4.post2
